# docs/swagger.yaml
definitions:
...
  user_dto.UserLoginRequest:
    properties:
      userAccount:
        description: Identity     string `json:"identity,omitempty"`
        type: string
      userPassword:
        type: string
    required:
    - userAccount
    - userPassword
    type: object
  user_dto.UserUpdateMyUserRequest:
    properties:
      userAvatar:
        type: string
      userName:
        type: string
      userProfile:
        type: string
    type: object
  user_vo.UserVO:
    properties:
      createTime:
        type: string
      identity:
        type: string
      updateTime:
        type: string
      userAvatar:
        type: string
      userName:
        type: string
      userProfile:
        type: string
      userRole:
        type: string
    type: object
  vo.BaseResponse-bool:
    properties:
      code:
        description: 响应码
        type: integer
      data:
        description: 接口，表示具体信息
        type: boolean
      message:
        description: 请求结果[发生错误则是错误信息，如果没有错误则是SUCCESS.Code的值]
        type: string
    type: object
...
  vo.BaseResponse-user_vo_UserVO:
    properties:
      code:
        description: 响应码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/user_vo.UserVO'
        description: 接口，表示具体信息
      message:
        description: 请求结果[发生错误则是错误信息，如果没有错误则是SUCCESS.Code的值]
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/user/login:
    post:
      parameters:
      - description: user login infos
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/user_dto.UserLoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BaseResponse-user_vo_UserVO'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 用户登陆
      tags:
      - 用户
  /api/user/logout:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BaseResponse-bool'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 用户退出登录
      tags:
      - 用户
  /api/user/my:
    post:
      parameters:
      - description: user login infos
        in: body
        name: updateInfo
        required: true
        schema:
          $ref: '#/definitions/user_dto.UserUpdateMyUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BaseResponse-bool'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 用户更新自己的数据
      tags:
      - 用户
...
swagger: "2.0"